+++ NewScene/Assets/Script/Other/HealthManager.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class HealthManager : MonoBehaviour
 {
-    public int maxHealth;
-    public int currentHealth;
+    public float maxHealth;
+    public float currentHealth;
     
     void Start()
     {
@@ -18,14 +18,14 @@ public class HealthManager : MonoBehaviour
        
     }
   
-    public void HurtPlayer(int damage)
+    public void HurtPlayer(float damage)
     {
         currentHealth -= damage;
         FindObjectOfType<Main_Player>().PlayerHP(currentHealth);
     }
 
 
-    public void HealPlayer(int healAmount)
+    public void HealPlayer(float healAmount)
     {
         currentHealth += healAmount;
